[refactor] Move `RemoteGridImageDataLoaderWithCache` to single file
